fastapiproject/requirements.txt
fastapi
uvicorn
tortoise-orm
pydantic
mysqlclient
aiomysql
aerich